+++ java/src/test/java/io/cucumber/messages/MessagesTest.java
@@ -37,6 +37,7 @@ public class MessagesTest {
                 null,
                 null,
                 null,
+                null,
                 null
         );
     }